Update runbot.py
Rewritten to support error handling of the listenchannel commands. Added descriptions to the hybrid commands to make them easier for users.

Added error handling for if a user is/isn't in the cfg.listeningUsers file.

Added error handling for if a user does not exist (Handles both Twitch and 7TV)